+++ pydapper/bigquery/google_bigquery_client.py
@@ -1,5 +1,7 @@
 from typing import TYPE_CHECKING
+from typing import ClassVar
 
+from pydapper.capabilities import AdapterCapability
 from pydapper.commands import Commands
 
 from ..utils import import_dbapi_module
@@ -9,6 +11,8 @@ if TYPE_CHECKING:
 
 
 class GoogleBigqueryClientCommands(Commands):
+    capabilities: ClassVar[frozenset[AdapterCapability]] = frozenset()
+
     @classmethod
     def connect(cls, parsed_dsn: "PydapperParseResult", **connect_kwargs) -> "Commands":
         google_bigquery_client = import_dbapi_module("google.cloud.bigquery.dbapi")